+++ lib/screens/auth/signUp/signUpOptionsScreen.dart
@@ -57,7 +57,7 @@ class SignUpOptionsScreen extends StatelessWidget {
             ),
             BuildSignUpOptionCard(
                 option: 'Instagram',
-                icon: "assets/icons/instagram.svg",
+                icon: "assets/icons/instagram1.svg",
                 press: () {}),
             SizedBox(
               height: getPropHeight(20),
